Refactor, api calls, working
