+++ src/components/mapLayers/HotSpots.tsx
@@ -1,13 +1,13 @@
 import { useLeafletContext } from "@react-leaflet/core";
 import { useEffect, useRef } from "react";
-import { Disaster } from "@/types/disaster";
 import L from "leaflet";
 
 import { getCountryColor } from "@/utils/countryColor";
+import { mapObject } from "@/server/client";
 
 type HotSpotsProps = {
-  disasterData: Disaster[];
-  setSelectedDisaster: (disaster: Disaster) => void;
+  disasterData: mapObject[];
+  setSelectedDisaster: (disaster: mapObject) => void;
 };
 
 export default function HotSpots({
