--- src/types/org.ts
@@ -1,5 +1,0 @@
-export type Org = {
-  name: string;
-  url: string;
-  donationsMatched: number;
-};
